Wireframe_v1.0
Created a wireframe based on the needs of our website. Primarily
focusing on using a large header to inform and draw in the user. The
Text Area will vary depending on content. Initial thoughts of this
website is that it will be a one page website with few secondary pages,
i.e biographies, blog posts and additional information on the project.