Can get all the tweets!!
Firstly, i cleaned up the code, was getting messy so i have a separate
file for functions called Handlers, a file for the Interface and the
main file.

So previously the tweets would only go so far and then loop back around.
Ive added in some code that will allow the program to automatically get
the next page of tweets. Its working really well now!